test(mw,sess): co-locate package tests into mw_test and sess_test
Move the mw.* and sess.* tests out of the root via_test package into
package mw_test / sess_test so coverage attributes to the package under
test (both previously read 0% — a measurement artifact, not missing
tests). mw now reports 93.6%, sess 84.6%. Also adds mw.CSP boundary
tests that exercise the middleware directly via httptest.NewRecorder.
